+++ qom/systems/SOMASystem.py
@@ -6,7 +6,7 @@
 __name__    = 'qom.systems.SOMASystem'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2021-08-15'
-__updated__ = '2022-04-24'
+__updated__ = '2022-07-26'
 
 # dependencies
 from decimal import Decimal
@@ -20,6 +20,8 @@ from ..solvers import ODESolver
 
 # module logger
 logger = logging.getLogger(__name__)
+
+# TODO: Implement integration in `get_lle_dynamics`.
 
 class SOMASystem(BaseSystem):
     """Class to interface an array of single-optomechanical systems.
@@ -66,6 +68,7 @@ class SOMASystem(BaseSystem):
         # update attributes
         self.required_funcs.update({
             'get_mode_amplitude_dynamics': ['get_ivc', 'get_mode_rates'],
+            'get_lle_dynamics': ['get_Ds', 'get_ivc', 'get_Ns', 'get_S', 'get_X'],
             'get_nlse_dynamics': ['get_ivc', 'get_op_d', 'get_op_n']
         })
         self.required_params.update({
@@ -77,8 +80,8 @@ class SOMASystem(BaseSystem):
            'mode_type': 'optical'
         })
 
-    def get_mode_amplitude_dynamics(self, solver_params: dict, plot: bool=False, plotter_params: dict=dict()):
-        """Method to obtain the dynamics of the optical mode amplitudes by solving the classical equations of motion.
+    def get_mode_intensity_dynamics(self, solver_params: dict, plot: bool=False, plotter_params: dict=dict()):
+        """Method to obtain the dynamics of the optical or mechanical mode amplitudes by solving the classical equations of motion.
 
         Requires predefined callables ``get_ivc`` and ``get_mode_rates``.
         
@@ -95,8 +98,8 @@ class SOMASystem(BaseSystem):
 
         Returns
         -------
-        amps : list
-            Required mode amplitudes at every point of time.
+        I : list
+            Required mode intensities at every point of time.
         T : list 
             Times at which the measures are calculated.
         """
@@ -114,14 +117,109 @@ class SOMASystem(BaseSystem):
         _range_max = int(solver_params.get('range_max', len(T)))
         N = list(range(_n))
 
-        # extract required modes
-        amps = [[np.linalg.norm(Modes[i][2 * j + (1 if _mode_type == 'mechanical' else 0)]) for j in range(_n)] for i in range(_range_min, _range_max)]
+        # extract required mode intensities
+        I = [[np.linalg.norm(Modes[i][2 * j + (1 if _mode_type == 'mechanical' else 0)]) for j in range(_n)] for i in range(_range_min, _range_max)]
 
         # if plot opted
         if plot:
-            self.plot_dynamics(plotter_params=plotter_params, V=amps, T=T, N=N)
+            self.plot_dynamics(plotter_params=plotter_params, V=I, T=T, N=N)
 
-        return amps, T[_range_min:_range_max], N
+        return I, T[_range_min:_range_max], N
+
+    def get_lle_dynamics(self, solver_params):
+        """Method to obtain the solutions of the Lugiato-Lefever equation (LLE) using the Split-step Fourier Method.
+
+        Requires predefined callables ``get_D``, ``get_ivc``, ``get_N``, ``get_S`` and ``get_X``. Also, either one of the functions ``get_betas`` and ``get_beta_rates`` should be defined.
+
+        Parameters
+        ----------
+        solver_params : dict
+            Parameters for the solver containing the keys "t_min", "t_max" and "t_dim" for the integration timescale.
+
+        Returns
+        -------
+        Modes : list
+            Calculated solutions.
+        X : list
+            List of X-axis values of the system.
+        T : list
+            Times at which the solutions are obtainted.
+        """
+
+        # validate required functions
+        assert self.validate_required_funcs(func_name='get_lle_dynamics', mode='verbose') is True, 'Missing required predefined functions'
+
+        # validate optional functions
+        assert getattr(self, 'get_beta_rates', None) is not None or getattr(self, 'get_betas', None) is not None, 'Either of the functions ``get_beta_rates`` or ``get_betas`` should be non-none'
+
+        # validate parameters
+        for key in ['t_min', 't_max', 't_dim']:
+            assert key in solver_params, 'Parameter ``params`` should contain key "{}" for looper parameters'.format(key)
+                
+        # extract frequently used variables
+        t_min = np.float_(solver_params['t_min'])
+        t_max = np.float_(solver_params['t_max'])
+        t_dim = int(solver_params['t_dim'])
+        # _method = solver_params.get('method', 'RK45')
+        _show_progress = solver_params.get('show_progress', False)
+
+        # calculate times
+        ts = np.linspace(t_min, t_max, t_dim)
+        # truncate values
+        _ss = (Decimal(str(t_max)) - Decimal(str(t_min))) / (t_dim - 1)
+        _decimals = - _ss.as_tuple().exponent
+        # round off and convert to list
+        T = np.around(ts, _decimals).tolist()
+        # steps size
+        t_ss = T[1] - T[0]
+
+        # extract parameters
+        iv, c = self.get_ivc()
+        modes = iv[:self.num_modes]
+        if len(c) > 4 * self.num_modes**2:
+            params = c[4 * self.num_modes**2:]
+        else:
+            params = c
+    
+        # initialize variables
+        Modes = list()
+        Modes.append([m for m in modes])
+        alphas = np.array(modes[::2])
+
+        for i in range(1, t_dim):
+            t = (i - 1) * t_ss + t_min
+            # update progress
+            if _show_progress:
+                self._update_progress(module_name=__name__, pos=i, dim=t_dim)
+
+            # apply nonlinearity
+            Ns = self.get_Ns(modes, params, t)
+            S = self.get_S(modes, params, t)
+            alpha_nls = np.exp(Ns * t_ss) * (alphas + S / Ns) - S / Ns
+
+            # apply dispersion
+            alpha_tildes = sf.fft(alpha_nls)
+            # solver = ODESolver(params=solver_params, func=lambda t, v, c: self.get_Ds(modes=v, params=c, t=t), iv=alpha_tildes, c=params, method=_method, cb_update=self.cb_update)
+            # alpha_tildes = solver.solve(T=[0, t_ss])[-1]
+            Ds = self.get_Ds(modes, params, t)
+            alpha_tildes = np.exp(Ds * t_ss) * alpha_tildes
+            alphas = sf.ifft(alpha_tildes)
+
+            # update lists
+            for j in range(len(alphas)):
+                modes[2 * j] = alphas[j]
+            betas = self.get_betas(modes, params, t)
+            for j in range(len(alphas)):
+                modes[2 * j + 1] = betas[j]
+            Modes.append([m for m in modes])
+
+        # display completion
+        if _show_progress:
+            logger.info('------------------Dynamics Obtained------------------\n')
+            if self.cb_update is not None:
+                self.cb_update(status='Dynamics Obtained', progress=None, reset=True)
+
+        return Modes, self.get_X(params), T
 
     def get_nlse_dynamics(self, solver_params):
         """Method to obtain the solutions of the Nonlinear Schrodinger Equation (NLSE) using the Split-step Fourier Method.